% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETAregions_labels.R
\docType{data}
\name{ETAregions_labels}
\alias{ETAregions_labels}
\title{ETAregions_labels shapefile data}
\format{
Simple feature collection
\describe{
\item{eta_region}{Number code representing ETA region, 1-6}
\item{eta_region_name}{Name of region}
\item{etaA_region_city}{City headquarters for ETA region}
\item{geometry}{Point coordinates for labels on ETA regions}
}
}
\source{
Coordinates for hex map are based on NPR's daily graphic template,
\url{https://github.com/nprapps/dailygraphics-templates/blob/master/state_grid_map/index.html}

Department of Labor, Employment and Training Administration
\url{https://www.dol.gov/agencies/eta}

Workforce Innovation and Opportunity Act
\url{https://www.dol.gov/agencies/eta/wioa/}

ETA regions
\url{https://www.dol.gov/agencies/eta/regions}
}
\usage{
ETAregions_labels
}
\description{
Point coordinates for the labels on ETA regions.
}
\details{
These are based on ETA (Department of Labor's Employment and Training Administration) regions for reporting WIOA (Workforce Innovation and Opportunity Act) performance.
}
\keyword{datasets}
