% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coordinates.R
\name{get_coordinates}
\alias{get_coordinates}
\title{Import hex map coordinates}
\usage{
get_coordinates(map = NULL, coords = "hexmap")
}
\arguments{
\item{map}{A \code{character} string specifying which map.
Current options are
\code{"usa50"} (50 US States only),
\code{"usa51"} (50 states and DC),
\code{"usa52"} (50 states, DC, and Puerto Rico),
\code{"usa53"} (50 states, DC, PUerto Rico and Virgin Islands),
\code{"usa56"} (50 states, DC, Puerto Rico, Virgin Islands, Guam,
American Samoa, and Marshall Islands),
\code{"usaETA"} (WIOA ETA geographies, same as statesTerr with the addition
of Palau), and
\code{"ETARegions"} (unified maps of WIOA ETA regions) .}

\item{coords}{A \code{character} string specifying \code{"hexmap"} for the
hexagon coordinates or \code{"labels"} for the label coordinates for the hex map}
}
\value{
sf object for map, either the hexagons (polygon) or labels (point).
}
\description{
Import hex map coordinates
}
\examples{
get_coordinates(map = "usa50", coords = "hexmap")
get_coordinates(map = "usa50", coords = "labels")
}
