% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statesTerr_labels.R
\docType{data}
\name{statesTerr_labels}
\alias{statesTerr_labels}
\title{statesTerr_labels shapefile data}
\format{
Simple feature collection
\describe{
\item{abbr_usps}{2-letter code used by the United States Postal Service}
\item{abbr_gpo}{abbreviations from US Government Printing Office}
\item{abbr_ap}{abbreviations from AP Stylebook}
\item{name}{Name of geographic area}
\item{fips}{2-digit FIPS code (Federal Information Processing Standard)}
\item{geo_type}{Type of region: state, federal district, territory, freely associated state}
\item{geometry}{point coordinates for the center of the hexagon}
}
}
\source{
Coordinates for hex map are based on NPR's daily graphic template,
\url{https://github.com/nprapps/dailygraphics-templates/blob/master/state_grid_map/index.html}

Abbreviations are pulled from List of U.S. state and territory abbreviations,
\url{https://en.wikipedia.org/wiki/List_of_U.S._state_and_territory_abbreviations}
}
\usage{
statesTerr_labels
}
\description{
Point coordinates for the center of the hexagon that includes:
\itemize{
\item 50 States
\item 1 federal district: District of Columbia
\item 5 territories: Puerto Rico, Virgin Islands, American Samoa, and the Northern Mariana Islands
}
}
\keyword{datasets}
