% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIOAETA_labels.R
\docType{data}
\name{WIOAETA_labels}
\alias{WIOAETA_labels}
\title{WIOAETA_labels shapefile data}
\format{
Simple feature collection
\describe{
\item{abbr_usps}{2-letter code used by the United States Postal Service}
\item{abbr_gpo}{abbreviations from US Government Printing Office}
\item{abbr_ap}{abbreviations from AP Stylebook}
\item{name}{Name of geographic area}
\item{fips}{2-digit FIPS code (Federal Information Processing Standard)}
\item{geo_type}{Type of region: state, federal district, territory, freely associated state}
\item{geometry}{point coordinates for the center of the hexagon}
}
}
\source{
Coordinates for hex map are based on NPR's daily graphic template,
\url{https://github.com/nprapps/dailygraphics-templates/blob/master/state_grid_map/index.html}

Abbreviations are pulled from List of U.S. state and territory abbreviations,
\url{https://en.wikipedia.org/wiki/List_of_U.S._state_and_territory_abbreviations}

Department of Labor, Employment and Training Administration
\url{https://www.dol.gov/agencies/eta}

Workforce Innovation and Opportunity Act
\url{https://www.dol.gov/agencies/eta/wioa/}

ETA regions
\url{https://www.dol.gov/agencies/eta/regions}
}
\usage{
WIOAETA_labels
}
\description{
Point coordinates for the center of the hexagon that includes:
\itemize{
\item 50 States
\item 1 federal district: District of Columbia
\item 5 territories: Puerto Rico, Virgin Islands, American Samoa, and the Northern Mariana Islands
\item 1 freely associated state: Palau
}
}
\details{
These are based on ETA (Department of Labor's Employment and Training Administration) regions for reporting WIOA (Workforce Innovation and Opportunity Act) performance.
}
\keyword{datasets}
